OTHERS/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-dotenv==1.0.1
httpx==0.27.0
Pillow==10.2.0
python-multipart==0.0.9
Jinja2==3.1.3
langchain-gigachat>=0.3.10
langchain-core>=0.1.27
psycopg2-binary==2.9.9
PyMuPDF==1.23.22